+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### 2.0.0: Major change in implementation
+  * Change the `Promise` class to be a monad and return a new promise on `then` and `catch`
+  * Add more tests to improve coverage and edge cases
+
 ### 1.0.5: Add ESLint
 
 ### 1.0.4: Patching dependency vulnerabilities